Fixed double-float confusion in wavefold.h. Was error on gcc-arm

+++ src/processors/wavefold.h
@@ -45,9 +45,9 @@ public:
 		//Generate fold tables
 		for (int i = 0; i < 512; i++) {
 			fold1[i] = map_value(i, 0.0f, 511.0f, -1.0f, 1.0f);
-			fold2[i] = map_value(fmod(i, 256), 0.0f, 255.0f, -1.0f, 1.0f);
-			fold3[i] = map_value(fmod(i, 170.6666666667), 0.0f, (512 / 3.0f - 1), -1.0f, 1.0f);
-			fold4[i] = map_value(fmod(i, 128), 0.0f, 127.0f, -1.0f, 1.0f);
+			fold2[i] = map_value(fmodf(i, 256.f), 0.0f, 255.0f, -1.0f, 1.0f);
+			fold3[i] = map_value(fmodf(i, 170.6666666667), 0.0f, (512.f / 3.0f - 1.f), -1.0f, 1.0f);
+			fold4[i] = map_value(fmodf(i, 128.f), 0.0f, 127.0f, -1.0f, 1.0f);
 		}
 	}
